Add the machine of states
